Trait ra_hir::db::HirDatabase

pub trait HirDatabase: DefDatabase + AstDatabase {
    fn expr_scopes(&self, key0: DefWithBody) -> Arc<ExprScopes>;

    fn infer(&self, key0: DefWithBody) -> Arc<InferenceResult>;

    fn type_for_def(&self, key0: TypableDef, key1: Namespace) -> Ty;

    fn type_for_field(&self, key0: StructField) -> Ty;

    fn callable_item_signature(&self, key0: CallableDef) -> FnSig;

    fn generic_predicates_for_param(
        &self, 
        key0: GenericDef, 
        key1: u32
    ) -> Arc<[GenericPredicate]>;

    fn generic_predicates(&self, key0: GenericDef) -> Arc<[GenericPredicate]>;

    fn generic_defaults(&self, key0: GenericDef) -> Substs;

    fn body_with_source_map(
        &self, 
        key0: DefWithBody
    ) -> (Arc<Body>, Arc<BodySourceMap>);

    fn body_hir(&self, key0: DefWithBody) -> Arc<Body>;

    fn impls_in_crate(&self, key0: Crate) -> Arc<CrateImplBlocks>;

    fn impls_for_trait(&self, key0: Crate, key1: Trait) -> Arc<[ImplBlock]>;

    fn trait_solver(&self, key0: Crate) -> TraitSolver;

    fn intern_type_ctor(&self, key0: TypeCtor) -> TypeCtorId;

    fn lookup_intern_type_ctor(&self, key0: TypeCtorId) -> TypeCtor;

    fn intern_impl(&self, key0: Impl) -> GlobalImplId;

    fn lookup_intern_impl(&self, key0: GlobalImplId) -> Impl;

    fn associated_ty_data(&self, key0: TypeId) -> Arc<AssociatedTyDatum>;

    fn trait_datum(&self, key0: Crate, key1: TraitId) -> Arc<TraitDatum>;

    fn struct_datum(&self, key0: Crate, key1: StructId) -> Arc<StructDatum>;

    fn impl_datum(&self, key0: Crate, key1: ImplId) -> Arc<ImplDatum>;

    fn trait_solve(
        &self, 
        key0: Crate, 
        key1: Canonical<InEnvironment<Obligation>>
    ) -> Option<Solution>;
}

Required methods

fn expr_scopes(&self, key0: DefWithBody) -> Arc<ExprScopes>

fn infer(&self, key0: DefWithBody) -> Arc<InferenceResult>

fn type_for_def(&self, key0: TypableDef, key1: Namespace) -> Ty

fn type_for_field(&self, key0: StructField) -> Ty

fn callable_item_signature(&self, key0: CallableDef) -> FnSig

fn generic_predicates_for_param(
&self,
key0: GenericDef,
key1: u32
) -> Arc<[GenericPredicate]>

fn generic_predicates(&self, key0: GenericDef) -> Arc<[GenericPredicate]>

fn generic_defaults(&self, key0: GenericDef) -> Substs

fn body_with_source_map(
&self,
key0: DefWithBody
) -> (Arc<Body>, Arc<BodySourceMap>)

fn body_hir(&self, key0: DefWithBody) -> Arc<Body>

fn impls_in_crate(&self, key0: Crate) -> Arc<CrateImplBlocks>

fn impls_for_trait(&self, key0: Crate, key1: Trait) -> Arc<[ImplBlock]>

fn trait_solver(&self, key0: Crate) -> TraitSolver

This provides the Chalk trait solver instance. Because Chalk always works from a specific crate, this query is keyed on the crate; and because Chalk does its own internal caching, the solver is wrapped in a Mutex and the query does an untracked read internally, to make sure the cached state is thrown away when input facts change.

fn intern_type_ctor(&self, key0: TypeCtor) -> TypeCtorId

fn lookup_intern_type_ctor(&self, key0: TypeCtorId) -> TypeCtor

fn intern_impl(&self, key0: Impl) -> GlobalImplId

fn lookup_intern_impl(&self, key0: GlobalImplId) -> Impl

fn associated_ty_data(&self, key0: TypeId) -> Arc<AssociatedTyDatum>

fn trait_datum(&self, key0: Crate, key1: TraitId) -> Arc<TraitDatum>

fn struct_datum(&self, key0: Crate, key1: StructId) -> Arc<StructDatum>

fn impl_datum(&self, key0: Crate, key1: ImplId) -> Arc<ImplDatum>

fn trait_solve(
&self,
key0: Crate,
key1: Canonical<InEnvironment<Obligation>>
) -> Option<Solution>

Loading content...