Trait ra_hir::diagnostics::Diagnostic

pub trait Diagnostic: Any + Send + Sync + Debug + 'static {
    fn message(&self) -> String;

    fn file(&self) -> HirFileId;

    fn syntax_node_ptr(&self) -> SyntaxNodePtr;

    fn as_any(&self) -> &(dyn Any + Send + 'static);

    fn highlight_range(&self) -> TextRange { ... }
}

Diagnostic defines hir API for errors and warnings.

It is used as a dyn object, which you can downcast to a concrete diagnostic. DiagnosticSink are structured, meaning that they include rich information which can be used by IDE to create fixes. DiagnosticSink are expressed in terms of macro-expanded syntax tree nodes (so, it's a bad idea to diagnostic in a salsa value).

Internally, various subsystems of hir produce diagnostics specific to a subsystem (typically, an enum), which are safe to store in salsa but do not include source locations. Such internal diagnostic are transformed into an instance of Diagnostic on demand.

Required methods

fn message(&self) -> String

fn file(&self) -> HirFileId

fn syntax_node_ptr(&self) -> SyntaxNodePtr

fn as_any(&self) -> &(dyn Any + Send + 'static)

Loading content...
Provided methods

fn highlight_range(&self) -> TextRange

Loading content...
Methods

impl dyn Diagnostic

pub fn syntax_node(&self, db: &impl HirDatabase) -> TreeArc<SyntaxNode>

pub fn downcast_ref<D: Diagnostic>(&self) -> Option<&D>