Enum ra_hir::Ty

pub enum Ty {
    Apply(ApplicationTy),
    Param {
        idx: u32,
        name: Name,
    },
    Bound(u32),
    Infer(InferTy),
    Unknown,
}

A type.

See also the TyKind enum in rustc (librustc/ty/sty.rs), which represents the same thing (but in a different way).

This should be cheap to clone.

Variants
Apply(ApplicationTy)

A nominal type with (maybe 0) type parameters. This might be a primitive type like bool, a struct, tuple, function pointer, reference or several other things.

Param

A type parameter; for example, T in `fn f(x: T) {}

Fields of Param

idx: u32

The index of the parameter (starting with parameters from the surrounding impl, then the current function).

name: Name

The name of the parameter, for displaying.

Bound(u32)

A bound type variable. Only used during trait resolution to represent Chalk variables.

Infer(InferTy)

A type variable used during type checking. Not to be confused with a type parameter.

Unknown

A placeholder for a type which could not be computed; this is propagated to avoid useless error messages. Doubles as a placeholder where type variables are inserted before type checking, since we want to try to infer a better type here anyway -- for the IDE use case, we want to try to infer as much as possible even in the presence of type errors.

Methods

impl Ty

pub fn autoderef<'a>(
self,
db: &'a impl HirDatabase
) -> impl Iterator<Item = Ty> + 'a

Iterates over the possible derefs of ty.

impl Ty

pub fn iterate_impl_items<T>(
self,
db: &impl HirDatabase,
krate: Crate,
callback: impl FnMut(ImplItem) -> Option<T>
) -> Option<T>

impl Ty

pub fn simple(ctor: TypeCtor) -> Ty

pub fn apply_one(ctor: TypeCtor, param: Ty) -> Ty

pub fn apply(ctor: TypeCtor, parameters: Substs) -> Ty

pub fn unit() -> Self

pub fn walk(&self, f: &mut impl FnMut(&Ty))

pub fn as_reference(&self) -> Option<(&Ty, Mutability)>

pub fn as_adt(&self) -> Option<(AdtDef, &Substs)>

pub fn as_tuple(&self) -> Option<&Substs>

pub fn as_callable(&self) -> Option<(CallableDef, &Substs)>

pub fn apply_substs(self, substs: Substs) -> Ty

If this is a type with type parameters (an ADT or function), replaces the Substs for these type parameters with the given ones. (So e.g. if self is Option<_> and the substs contain u32, we'll have Option<u32> afterwards.)

pub fn subst(self, substs: &Substs) -> Ty

Replaces type parameters in this type using the given Substs. (So e.g. if self is &[T], where type parameter T has index 0, and the Substs contain u32 at index 0, we'll have &[u32] afterwards.)

pub fn subst_bound_vars(self, substs: &Substs) -> Ty

Substitutes Ty::Bound vars (as opposed to type parameters).