Enum ra_hir::TypeCtor

pub enum TypeCtor {
    Bool,
    Char,
    Int(UncertainIntTy),
    Float(UncertainFloatTy),
    Adt(Adt),
    Str,
    Slice,
    Array,
    RawPtr(Mutability),
    Ref(Mutability),
    FnDef(CallableDef),
    FnPtr {
        num_args: u16,
    },
    Never,
    Tuple {
        cardinality: u16,
    },
    AssociatedType(TypeAlias),
    Closure {
        def: DefWithBody,
        expr: ExprId,
    },
}

A type constructor or type name: this might be something like the primitive type bool, a struct like Vec, or things like function pointers or tuples.

Variants
Bool

The primitive boolean type. Written as bool.

Char

The primitive character type; holds a Unicode scalar value (a non-surrogate code point). Written as char.

Int(UncertainIntTy)

A primitive integer type. For example, i32.

Float(UncertainFloatTy)

A primitive floating-point type. For example, f64.

Adt(Adt)

Structures, enumerations and unions.

Str

The pointee of a string slice. Written as str.

Slice

The pointee of an array slice. Written as [T].

Array

An array with the given length. Written as [T; n].

RawPtr(Mutability)

A raw pointer. Written as *mut T or *const T

Ref(Mutability)

A reference; a pointer with an associated lifetime. Written as &'a mut T or &'a T.

FnDef(CallableDef)

The anonymous type of a function declaration/definition. Each function has a unique type, which is output (for a function named foo returning an i32) as fn() -> i32 {foo}.

This includes tuple struct / enum variant constructors as well.

For example the type of bar here:

fn foo() -> i32 { 1 }
let bar = foo; // bar: fn() -> i32 {foo}

FnPtr

A pointer to a function. Written as fn() -> i32.

For example the type of bar here:

fn foo() -> i32 { 1 }
let bar: fn() -> i32 = foo;

Fields of FnPtr

num_args: u16
Never

The never type !.

Tuple

A tuple type. For example, (i32, bool).

Fields of Tuple

cardinality: u16
AssociatedType(TypeAlias)

Represents an associated item like Iterator::Item. This is used when we have tried to normalize a projection like T::Item but couldn't find a better representation. In that case, we generate an application type like (Iterator::Item)<T>.

Closure

The type of a specific closure.

The closure signature is stored in a FnPtr type in the first type parameter.

Fields of Closure

def: DefWithBodyexpr: ExprId

Methods

impl TypeCtor

pub fn num_ty_params(self, db: &impl HirDatabase) -> usize

pub fn krate(self, db: &impl HirDatabase) -> Option<Crate>

pub fn as_generic_def(self) -> Option<GenericDef>