Struct ra_hir::Semantics

pub struct Semantics<'db, DB> {
    pub db: &'db DB,
    // some fields omitted
}

Primary API to get semantic information, like types, from syntax trees.

Fields
db: &'db DB
Methods

impl<'db, DB: HirDatabase> Semantics<'db, DB>

pub fn new(db: &DB) -> Semantics<DB>

pub fn parse(&self, file_id: FileId) -> SourceFile

pub fn expand(&self, macro_call: &MacroCall) -> Option<SyntaxNode>

pub fn expand_hypothetical(
&self,
actual_macro_call: &MacroCall,
hypothetical_args: &TokenTree,
token_to_map: SyntaxToken
) -> Option<(SyntaxNode, SyntaxToken)>

pub fn descend_into_macros(&self, token: SyntaxToken) -> SyntaxToken

pub fn descend_node_at_offset<N: AstNode>(
&self,
node: &SyntaxNode,
offset: TextSize
) -> Option<N>

pub fn original_range(&self, node: &SyntaxNode) -> FileRange

pub fn diagnostics_range(&self, diagnostics: &dyn Diagnostic) -> FileRange

pub fn ancestors_with_macros(
&self,
node: SyntaxNode
) -> impl Iterator<Item = SyntaxNode> + '_

pub fn ancestors_at_offset_with_macros(
&self,
node: &SyntaxNode,
offset: TextSize
) -> impl Iterator<Item = SyntaxNode> + '_

pub fn find_node_at_offset_with_macros<N: AstNode>(
&self,
node: &SyntaxNode,
offset: TextSize
) -> Option<N>

Find a AstNode by offset inside SyntaxNode, if it is inside Macrofile, search up until it is of the target AstNode type

pub fn find_node_at_offset_with_descend<N: AstNode>(
&self,
node: &SyntaxNode,
offset: TextSize
) -> Option<N>

Find a AstNode by offset inside SyntaxNode, if it is inside MacroCall, descend it and find again

pub fn type_of_expr(&self, expr: &Expr) -> Option<Type>

pub fn type_of_pat(&self, pat: &Pat) -> Option<Type>

pub fn resolve_method_call(&self, call: &MethodCallExpr) -> Option<Function>

pub fn resolve_field(&self, field: &FieldExpr) -> Option<Field>

pub fn resolve_record_field(
&self,
field: &RecordField
) -> Option<(Field, Option<Local>)>

pub fn resolve_record_field_pat(&self, field: &RecordFieldPat) -> Option<Field>

pub fn resolve_macro_call(&self, macro_call: &MacroCall) -> Option<MacroDef>

pub fn resolve_path(&self, path: &Path) -> Option<PathResolution>

pub fn resolve_bind_pat_to_const(&self, pat: &BindPat) -> Option<ModuleDef>

pub fn record_literal_missing_fields(
&self,
literal: &RecordLit
) -> Vec<(Field, Type)>

pub fn record_pattern_missing_fields(
&self,
pattern: &RecordPat
) -> Vec<(Field, Type)>

pub fn to_def<T: ToDef>(&self, src: &T) -> Option<T::Def>

pub fn to_module_def(&self, file: FileId) -> Option<Module>

pub fn scope(&self, node: &SyntaxNode) -> SemanticsScope<'db, DB>

pub fn scope_at_offset(
&self,
node: &SyntaxNode,
offset: TextSize
) -> SemanticsScope<'db, DB>

pub fn scope_for_def(&self, def: Trait) -> SemanticsScope<'db, DB>

pub fn assert_contains_node(&self, node: &SyntaxNode)