Trait ra_hir::TypeWalk

pub trait TypeWalk {
    fn walk(&self, f: &mut impl FnMut(&Ty));

    fn walk_mut_binders(
        &mut self, 
        f: &mut impl FnMut(&mut Ty, usize), 
        binders: usize
    );

    fn walk_mut(&mut self, f: &mut impl FnMut(&mut Ty)) { ... }

    fn fold(self, f: &mut impl FnMut(Ty) -> Ty) -> Self { ... }

    fn subst(self, substs: &Substs) -> Self { ... }

    fn subst_bound_vars(self, substs: &Substs) -> Self { ... }

    fn shift_bound_vars(self, n: i32) -> Self { ... }
}

This allows walking structures that contain types to do something with those types, similar to Chalk's Fold trait.

Required methods

fn walk(&self, f: &mut impl FnMut(&Ty))

fn walk_mut_binders(
&mut self,
f: &mut impl FnMut(&mut Ty, usize),
binders: usize
)

Walk the type, counting entered binders.

Ty::Bound variables use DeBruijn indexing, which means that 0 refers to the innermost binder, 1 to the next, etc.. So when we want to substitute a certain bound variable, we can't just walk the whole type and blindly replace each instance of a certain index; when we 'enter' things that introduce new bound variables, we have to keep track of that. Currently, the only thing that introduces bound variables on our side are Ty::Dyn and Ty::Opaque, which each introduce a bound variable for the self type.

Loading content...
Provided methods

fn walk_mut(&mut self, f: &mut impl FnMut(&mut Ty))

fn fold(self, f: &mut impl FnMut(Ty) -> Ty) -> Self

fn subst(self, substs: &Substs) -> Self

Replaces type parameters in this type using the given Substs. (So e.g. if self is &[T], where type parameter T has index 0, and the Substs contain u32 at index 0, we'll have &[u32] afterwards.)

fn subst_bound_vars(self, substs: &Substs) -> Self

Substitutes Ty::Bound vars (as opposed to type parameters).

fn shift_bound_vars(self, n: i32) -> Self

Shifts up Ty::Bound vars by n.

Loading content...