Crate ra_hir

HIR (previously known as descriptors) provides a high-level object oriented access to Rust code.

The principal difference between HIR and syntax trees is that HIR is bound to a particular crate instance. That is, it has cfg flags and features applied. So, the relation between syntax and HIR is many-to-one.

Re-exports

pub use self::source_binder::SourceAnalyzer;

pub use self::source_binder::PathResolution;

pub use self::source_binder::ScopeEntryWithSyntax;

pub use self::source_binder::MacroByExampleDef;

Modules

db

diagnostics

mock

source_binder

Macros

crate_graph

Structs

ApplicationTy
A nominal type with (maybe 0) type parameters. This might be a primitive type like bool, a struct, tuple, function pointer, reference or several other things.

AstIdMap
Maps items' SyntaxNodes to ErasedFileAstIds and back.

Const

ConstSignature
The declared signature of a const.

Crate
hir::Crate describes a single crate. It's the main interface with which a crate's dependencies interact. Mostly, it should be just a proxy for the root module.

CrateDependency

Documentation
Holds documentation

Enum

EnumVariant

ErasedFileAstId

ExprScopes

FnSignature
The declared signature of a function.

Function

GenericParam
Data about a generic parameter (to a function, struct, impl, ...).

GenericParams
Data about the generic parameters of a function, struct, impl, etc.

HirFileId
hir makes heavy use of ids: integer (u32) handlers to various things. You can think of id as a pointer (but without a lifetime) or a file descriptor (but for hir objects).

ImplBlock

ImportId

MacroCallId
MacroCallId identifies a particular macro invocation, like println!("Hello, {}", world).

MacroCallLoc

MacroDef

MacroDefId

Module

Name
Name is a wrapper around string, which is used in hir for both references and declarations. In theory, names should also carry hygiene info, but we are not there yet!

Path

PerNs

Static

Struct

StructField

Substs
A list of substitutions for generic parameters.

Trait

TraitRef
A trait with type parameters. This includes the Self, so this represents a concrete type implementing the trait. Name to be bikeshedded: TraitBound? TraitImplements?

TypeAlias

Union

Enums

AdtDef

BuiltinType

CallableDef

Container

DefWithBody
The defs which have a body.

Either

FieldSource

ImplItem

ModuleDef
The defs which can be visible in the module.

ModuleSource

Namespace

PathKind

Resolution

Ty
A type.

TypeCtor
A type constructor or type name: this might be something like the primitive type bool, a struct like Vec, or things like function pointers or tuples.

Traits

Docs

HasGenericParams

HirDisplay