Trait ra_hir::db::HirDatabase

pub trait HirDatabase: DefDatabase {
    fn expr_scopes(&self, key0: DefWithBody) -> Arc<ExprScopes>;

    fn infer(&self, key0: DefWithBody) -> Arc<InferenceResult>;

    fn type_for_def(&self, key0: TypableDef, key1: Namespace) -> Ty;

    fn type_for_field(&self, key0: StructField) -> Ty;

    fn callable_item_signature(&self, key0: CallableDef) -> FnSig;

    fn body_with_source_map(
        &self, 
        key0: DefWithBody
    ) -> (Arc<Body>, Arc<BodySourceMap>);

    fn body_hir(&self, key0: DefWithBody) -> Arc<Body>;

    fn impls_in_crate(&self, key0: Crate) -> Arc<CrateImplBlocks>;

    fn impls_for_trait(&self, key0: Crate, key1: Trait) -> Arc<[ImplBlock]>;

    fn solver(&self, key0: Crate) -> Arc<Mutex<Solver>>;

    fn implements(
        &self, 
        key0: Crate, 
        key1: Canonical<TraitRef>
    ) -> Option<Solution>;
}

Required methods

fn expr_scopes(&self, key0: DefWithBody) -> Arc<ExprScopes>

fn infer(&self, key0: DefWithBody) -> Arc<InferenceResult>

fn type_for_def(&self, key0: TypableDef, key1: Namespace) -> Ty

fn type_for_field(&self, key0: StructField) -> Ty

fn callable_item_signature(&self, key0: CallableDef) -> FnSig

fn body_with_source_map(
&self,
key0: DefWithBody
) -> (Arc<Body>, Arc<BodySourceMap>)

fn body_hir(&self, key0: DefWithBody) -> Arc<Body>

fn impls_in_crate(&self, key0: Crate) -> Arc<CrateImplBlocks>

fn impls_for_trait(&self, key0: Crate, key1: Trait) -> Arc<[ImplBlock]>

fn solver(&self, key0: Crate) -> Arc<Mutex<Solver>>

This provides the Chalk trait solver instance. Because Chalk always works from a specific crate, this query is keyed on the crate; and because Chalk does its own internal caching, the solver is wrapped in a Mutex and the query is marked volatile, to make sure the cached state is thrown away when input facts change.

fn implements(&self, key0: Crate, key1: Canonical<TraitRef>) -> Option<Solution>

Loading content...