Trait ra_hir::TypeWalk

pub trait TypeWalk {
    fn walk(&self, f: &mut impl FnMut(&Ty));

    fn walk_mut(&mut self, f: &mut impl FnMut(&mut Ty));

    fn fold(self, f: &mut impl FnMut(Ty) -> Ty) -> Self
    where
        Self: Sized,
    { ... }

    fn subst(self, substs: &Substs) -> Self
    where
        Self: Sized,
    { ... }

    fn subst_bound_vars(self, substs: &Substs) -> Self
    where
        Self: Sized,
    { ... }

    fn shift_bound_vars(self, n: i32) -> Self
    where
        Self: Sized,
    { ... }
}

This allows walking structures that contain types to do something with those types, similar to Chalk's Fold trait.

Required methods

fn walk(&self, f: &mut impl FnMut(&Ty))

fn walk_mut(&mut self, f: &mut impl FnMut(&mut Ty))

Loading content...
Provided methods

fn fold(self, f: &mut impl FnMut(Ty) -> Ty) -> Self where
Self: Sized,

fn subst(self, substs: &Substs) -> Self where
Self: Sized,

Replaces type parameters in this type using the given Substs. (So e.g. if self is &[T], where type parameter T has index 0, and the Substs contain u32 at index 0, we'll have &[u32] afterwards.)

fn subst_bound_vars(self, substs: &Substs) -> Self where
Self: Sized,

Substitutes Ty::Bound vars (as opposed to type parameters).

fn shift_bound_vars(self, n: i32) -> Self where
Self: Sized,

Shifts up Ty::Bound vars by n.

Loading content...